fix: dapp swap ui and metrics issues
