+++ shared/modules/dapp-swap-comparison/dapp-swap-comparison-utils.test.ts
@@ -6,6 +6,7 @@ import {
   getDataFromSwap,
   getBalanceChangeFromSimulationData,
   parseTransactionData,
+  checkValidSingleOrBatchTransaction,
 } from './dapp-swap-comparison-utils';
 
 const MOCK_QUOTES = [
@@ -193,6 +194,33 @@ describe('dapp-swap utils', () => {
         tokenBalanceChanges: [],
       });
       expect(result).toBe('0');
+    });
+  });
+
+  describe('checkValidSingleOrBatchTransaction', () => {
+    it('throws error for single transaction with invalid data', () => {
+      expect(() =>
+        checkValidSingleOrBatchTransaction([{ data: '0x' }]),
+      ).toThrow();
+    });
+
+    it('returns undefined for valid batch transaction', () => {
+      const result = checkValidSingleOrBatchTransaction([
+        { data: '0x3593564c0000' },
+        { data: '0x87517c450000' },
+      ]);
+      expect(result).toStrictEqual(undefined);
+    });
+
+    it('throws error if number of nested transactions is greater than 3', () => {
+      expect(() =>
+        checkValidSingleOrBatchTransaction([
+          { data: '0x' },
+          { data: '0x' },
+          { data: '0x' },
+          { data: '0x' },
+        ]),
+      ).toThrow();
     });
   });
 });
