+++ ui/components/ui/tabs/tabs.types.ts
@@ -37,6 +37,8 @@ export type TabsProps<TKey extends string = string> = Omit<
 > & {
   /** Key of the tab that should be active by default */
   defaultActiveTabKey?: TKey;
+  /** Key of the tab that should be active */
+  activeTabKey?: TKey;
   /** Callback called when a tab is clicked */
   onTabClick?: (tabKey: TKey) => void;
   /** Tab components to render */
