+++ ui/pages/confirmations/hooks/transactions/dapp-swap-comparison/useDappSwapComparisonInfo.ts
@@ -8,6 +8,7 @@ import { useSelector } from 'react-redux';
 
 import { captureException } from '../../../../../../shared/lib/sentry';
 import {
+  checkValidSingleOrBatchTransaction,
   getDataFromSwap,
   getBestQuote,
   getBalanceChangeFromSimulationData,
@@ -16,7 +17,6 @@ import {
 import { TokenStandAndDetails } from '../../../../../store/actions';
 import { getRemoteFeatureFlags } from '../../../../../selectors/remote-feature-flags';
 import { ConfirmMetamaskState } from '../../../types/confirm';
-import { checkValidSingleOrBatchTransaction } from '../../../utils';
 import { getTokenValueFromRecord } from '../../../utils/token';
 import { selectDappSwapComparisonData } from '../../../selectors/confirm';
 import { useConfirmContext } from '../../../context/confirm';
@@ -92,6 +92,7 @@ export function useDappSwapComparisonInfo() {
       );
       if (result.quotesInput) {
         updateRequestDetectionLatency();
+        captureDappSwapComparisonLoading(dataCommands);
       }
       return { ...result, commands: dataCommands };
     } catch (error) {
@@ -108,17 +109,12 @@ export function useDappSwapComparisonInfo() {
     }
   }, [
     captureDappSwapComparisonFailed,
+    captureDappSwapComparisonLoading,
     chainId,
     data,
     nestedTransactions,
     updateRequestDetectionLatency,
   ]);
-
-  useEffect(() => {
-    if (commands) {
-      captureDappSwapComparisonLoading(commands);
-    }
-  }, [captureDappSwapComparisonLoading, commands]);
 
   const {
     fiatRates,
