+++ ui/pages/confirmations/utils/confirm.test.ts
@@ -14,7 +14,6 @@ import {
   isValidASCIIURL,
   toPunycodeURL,
   stripProtocol,
-  checkValidSingleOrBatchTransaction,
 } from './confirm';
 
 const typedDataMsg =
@@ -143,33 +142,6 @@ describe('confirm util', () => {
       expect(stripProtocol('http://localhost:8545')).toStrictEqual(
         'localhost:8545',
       );
-    });
-  });
-
-  describe('checkValidSingleOrBatchTransaction', () => {
-    it('throws error for single transaction with invalid data', () => {
-      expect(() =>
-        checkValidSingleOrBatchTransaction([{ data: '0x' }]),
-      ).toThrow();
-    });
-
-    it('returns undefined for valid batch transaction', () => {
-      const result = checkValidSingleOrBatchTransaction([
-        { data: '0x3593564c0000' },
-        { data: '0x87517c450000' },
-      ]);
-      expect(result).toStrictEqual(undefined);
-    });
-
-    it('throws error if number of nested transactions is greater than 3', () => {
-      expect(() =>
-        checkValidSingleOrBatchTransaction([
-          { data: '0x' },
-          { data: '0x' },
-          { data: '0x' },
-          { data: '0x' },
-        ]),
-      ).toThrow();
     });
   });
 });
